twee structural changes
Modified the tables to be based on RadTableViewControllers,
which made it easier to manage styles and to switch between
displaying Twitter users and the Twee intro information.

Now the main "TwitterMob" screen is a table view. When we
aren't displaying Twitter users, the table view is configured
with a plist generated in routes.nu. When we are displaying
Twitter users, the table view has a single (large) cell that
contains the animated circles.

I believe that we can dynamically adjust the height of the table
cell to keep it from getting too much larger than our content,
but this is not done.